# server/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SummonScroll API",
    "version": "1.0.0",
    "description": "Production-ready API for SummonScroll - a gamified habit tracking application with monster collection mechanics",
    "contact": {
      "name": "SummonScroll Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/api/banners": {
      "get": {
        "tags": ["Banners"],
        "summary": "List active banners",
        "description": "Returns all active gacha banners. Cached for 60 seconds.",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Banners retrieved successfully",
            "headers": {
              "Cache-Control": {
                "schema": {"type": "string"},
                "description": "public, max-age=60, stale-while-revalidate=30"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedBannersResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/banners/{id}": {
      "get": {
        "tags": ["Banners"],
        "summary": "Get banner by ID",
        "description": "Returns detailed information about a specific banner",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "string", "format": "uuid"},
            "description": "Banner UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Banner retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Banner"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Banner not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/banners/{id}/pull": {
      "post": {
        "tags": ["Banners"],
        "summary": "Pull from banner",
        "description": "Performs a gacha pull (1x or 10x) from the specified banner. Rate limited to 10 pulls per minute.",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "string", "format": "uuid"},
            "description": "Banner UUID"
          },
          {
            "name": "X-RateLimit-Limit",
            "in": "header",
            "schema": {"type": "integer"},
            "description": "Rate limit: 10 pulls per minute"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "count": {
                    "type": "integer",
                    "enum": [1, 10],
                    "default": 1,
                    "description": "Number of pulls (1 or 10)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pull successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PullResult"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid pull count",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "402": {
            "description": "Insufficient currency",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "Insufficient spiritCrystals"
                }
              }
            }
          },
          "404": {
            "description": "Banner not found or inactive",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "Pull rate limit exceeded"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from /api/auth/login or /api/auth/register"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message"
          },
          "errors": {
            "type": "object",
            "description": "Validation errors (optional)"
          }
        }
      },
...
      "Banner": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "name": {"type": "string"},
          "description": {"type": "string"},
          "bannerType": {
            "type": "string",
            "enum": ["standard", "featured", "streak", "pact_seal", "event"]
          },
          "pullCost": {"type": "integer"},
          "pullCurrency": {
            "type": "string",
            "enum": ["spiritCrystals", "voidShards", "pactSeals"]
          },
          "isActive": {"type": "boolean"},
          "startsAt": {"type": "string", "format": "date-time"},
          "endsAt": {"type": "string", "format": "date-time", "nullable": true},
          "realmId": {"type": "string", "format": "uuid", "nullable": true},
          "realm": {
            "$ref": "#/components/schemas/Realm"
          }
        }
      },
      "Realm": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "name": {"type": "string"},
          "number": {"type": "integer"},
          "description": {"type": "string"}
        }
      },
      "Monster": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "name": {"type": "string"},
          "description": {"type": "string"},
          "rarity": {
            "type": "string",
            "enum": ["common", "uncommon", "rare", "elite", "epic", "legendary", "mythic", "ex"]
          },
          "element": {"type": "string"},
          "role": {"type": "string"},
          "baseHp": {"type": "integer"},
          "baseAtk": {"type": "integer"},
          "baseDef": {"type": "integer"},
          "baseSpeed": {"type": "integer"},
          "isEx": {"type": "boolean"},
          "bannerType": {"type": "string"},
          "artworkUrl": {"type": "string"},
          "iconUrl": {"type": "string"},
          "realmId": {"type": "string", "format": "uuid"},
          "realm": {
            "$ref": "#/components/schemas/Realm"
          },
          "skills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Skill"
            }
          }
        }
      },
      "Skill": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "name": {"type": "string"},
          "description": {"type": "string"},
          "cooldown": {"type": "integer"},
          "manaCost": {"type": "integer"}
        }
      },
...
      "PullResult": {
        "type": "object",
        "properties": {
          "monster": {
            "$ref": "#/components/schemas/Monster"
          },
          "isNew": {
            "type": "boolean",
            "description": "True if this is the first time obtaining this monster"
          },
          "isDuplicate": {
            "type": "boolean",
            "description": "True if user already owns this monster"
          },
          "transcendenceStoneGranted": {
            "type": "boolean",
            "description": "True if a transcendence stone was granted (EX monster duplicate)"
          }
        }
      },
...
      "PaginatedBannersResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Banner"
            }
          },
          "total": {"type": "integer"},
          "page": {"type": "integer"},
          "pageSize": {"type": "integer"},
          "hasMore": {"type": "boolean"}
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required or token invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "message": "Unauthorized"
            }
          }
        }
      },
...
    }
  }
}